refactor: update import paths for estimation test data map factory
